+++ CHANGELOG.md
@@ -28,6 +28,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Updated `@cumulus/sftp-client` and `@cumulus/ingest/SftpProviderClient` to support both regular and fastDownload.
   - Added sftp support to FakeProvider
   - Added sftp integration test
+- **CUMULUS-3919**
+  - Added terraform variables `disableSsl` and `rejectUnauthorized` to `tf-modules/cumulus-rds-tf` module.
 
 ### Changed